Introduce Native Flutter Splash Screen
The process was done using flutter_native_splash package.

Flutter application is shipped with a white splash screen which lasts about 3 seconds. Here the native screen was edited to show the desired image
set locally. The splash was constructed was Android only. So android was set to true in pubspec.yaml.

After setting up the assets 3 commands are to be run to make a clean build of the splash screen.
Commands :
- flutter clean
- flutter pub get
- flutter pub run flutter_native_splash:create

Note : 3 Commands can also be executed at the same time using the "&&" operator.

Signed-off-by: sajidshahriar72543 <sazidshahriar39@gmail.com>